attendance_website/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.0
django-widget-tweaks==1.4.8
numpy==1.19.3
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
